+++ src/main/java/org/realityforge/replicant/server/transport/EntityMessageAccumulator.java
@@ -4,6 +4,8 @@ import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.Map;
 import java.util.Map.Entry;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 import org.realityforge.replicant.server.EntityMessage;
 
 /**
@@ -12,38 +14,42 @@ import org.realityforge.replicant.server.EntityMessage;
  */
 public final class EntityMessageAccumulator
 {
-  private final Map<PacketQueue, LinkedList<EntityMessage>> _changeSets = new HashMap<>();
+  private final Map<ReplicantSession, LinkedList<EntityMessage>> _changeSets = new HashMap<>();
 
   /**
    * Add a message destined for a particular packet queue.
    *
-   * @param queue the queue.
+   * @param session the session.
    * @param message the message.
    */
-  public void addEntityMessage( final PacketQueue queue, final EntityMessage message )
+  public void addEntityMessage( final ReplicantSession session, final EntityMessage message )
   {
-    getChangeSet( queue ).add( message );
+    getChangeSet( session ).add( message );
   }
 
   /**
-   * Complete the collection of messages and forward them to the client.
+   * Complete the collection of messages and forward them to the clients.
+   *
+   * @param sessionID the session that initiated the changes.
+   * @param jobID the opaque identifier indicating the job that caused the changes.
    */
-  public void complete()
+  public void complete( @Nullable final String sessionID, @Nullable final String jobID )
   {
-    for ( final Entry<PacketQueue, LinkedList<EntityMessage>> entry : _changeSets.entrySet() )
+    for ( final Entry<ReplicantSession, LinkedList<EntityMessage>> entry : _changeSets.entrySet() )
     {
-      entry.getKey().addPacket( entry.getValue() );
+      final ReplicantSession session = entry.getKey();
+      session.getQueue().addPacket( session.getSessionID().equals( sessionID ) ? jobID : null, entry.getValue() );
     }
     _changeSets.clear();
   }
 
-  private LinkedList<EntityMessage> getChangeSet( final PacketQueue queue )
+  private LinkedList<EntityMessage> getChangeSet( @Nonnull final ReplicantSession session )
   {
-    LinkedList<EntityMessage> clientChangeSet = _changeSets.get( queue );
+    LinkedList<EntityMessage> clientChangeSet = _changeSets.get( session );
     if ( null == clientChangeSet )
     {
       clientChangeSet = new LinkedList<>();
-      _changeSets.put( queue, clientChangeSet );
+      _changeSets.put( session, clientChangeSet );
     }
     return clientChangeSet;
   }