+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.5.5:
+* Rework the transport system so that each packet can record the job id which
+  generated the change set. This is only done when the packet is for the session
+  which initiated the job.
 * Introduce ReplicantSession an base class from which sessions should extend.
 * Rework AbstractReplicationInterceptor so that it retrieves the session and job
   context information from the ReplicantContextHolder and passes it along to the